requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
email-validator==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.5.4
